Refactor: 로그인 로직, 유효성 검사 세분화 (#14)
