+++ src/shared/ui/icon/registry.ts
@@ -1,7 +1,9 @@
 import Arrow from '@/shared/assets/splash/icon-arrow.svg?react';
+import X from '@/shared/assets/icon-x.svg?react';
 
 export const iconRegistry = {
   arrow: Arrow,
+  x: X,
 } as const;
 
 export type IconName = keyof typeof iconRegistry;
