+++ src/shared/utils/validationSchemas.ts
@@ -2,8 +2,8 @@ import { z } from 'zod';
 
 // 로그인 폼 유효성 검사 스키마
 export const loginSchema = z.object({
-  email: z.string().min(1, '이메일을 입력해주세요').email('이메일 형식이 올바르지 않습니다.'),
-  password: z.string().min(1, '비밀번호를 입력해주세요'),
+  email: z.string().trim().email('이메일 형식이 올바르지 않습니다.'),
+  password: z.string(),
 });
 
 export type LoginFormData = z.infer<typeof loginSchema>;
